+++ packages/web_benchmarks/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0+10
+
+* Ensure the benchmark client reloads with the proper `initialPage`.
+* Migrates benchmark recorder away from deprecated `js_util` APIs.
+
 ## 0.1.0+9
 
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.